+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Perf
 
-### 2026-04-07
+### 2026-05-08
 
 - Replace synchronous disk I/O with async operations in snap sync [#6113](https://github.com/lambdaclass/ethrex/pull/6113)
+
+### 2026-04-27
+
+- Reduce peak disk usage during snap sync by moving SST files into the temp DB instead of copying [#6532](https://github.com/lambdaclass/ethrex/pull/6532)
 
 ### 2026-03-30